+++ blog-comment/src/main/java/com/mrqinzh/comment/mapper/CommentMapper.java
@@ -22,13 +22,13 @@ public interface CommentMapper extends BaseMapper<Comment> {
      * @param idType id类型： userId、articleId、commentId
      * @param id id值
      */
-    List<Comment> getById(@Param("idType") String idType, @Param("id") Integer id);
+    List<Comment> getById(@Param("idType") String idType, @Param("id") Long id);
 
     /**
      * 根据 id 删除评论信息
      * @param idType id类型： userId、articleId、commentId
      */
-    Boolean deleteByTypeId(@Param("idType") String idType, @Param("id") Integer id);
+    Boolean deleteByTypeId(@Param("idType") String idType, @Param("id") Long id);
 
     List<Comment> list(CommentPageDTO commentPageVo);
 }
